BUG:347798 Don't truncate fractional brush sizes on eraser switch
Patch by Alexey Elnatanov. Thanks!
CCMAIL: invalpoint@gmail.com

+++ krita/ui/kis_paintop_box.h
@@ -197,8 +197,8 @@ private:
     QString             m_currCompositeOpID;
     KisNodeWSP          m_previousNode;
 
-    int normalBrushSize; // when toggling between eraser mode
-    int eraserBrushSize;
+    qreal normalBrushSize; // when toggling between eraser mode
+    qreal eraserBrushSize;
 
     KisAction* m_hMirrorAction;
     KisAction* m_vMirrorAction;
